Level 3

+++ src/Check/Autoload_Options_Size.php
@@ -43,6 +43,6 @@ class Autoload_Options_Size extends Check {
 	private static function format_bytes( $size, $precision = 2 ) {
 		$base     = log( $size, 1024 );
 		$suffixes = array( '', 'kb', 'mb', 'g', 't' );
-		return round( pow( 1024, $base - floor( $base ) ), $precision ) . $suffixes[ floor( $base ) ];
+		return round( pow( 1024, $base - floor( $base ) ), $precision ) . $suffixes[ (int) floor( $base ) ];
 	}
 }
